protocol/vm/vmutil: move under vm package
Just organization to clarify that the vmutil package is for vm1
programs.
