+++ protocol/validation/block_test.go
@@ -7,7 +7,7 @@ import (
 	"chain/protocol/bc"
 	"chain/protocol/bc/legacy"
 	"chain/protocol/vm"
-	"chain/protocol/vmutil"
+	"chain/protocol/vm/vmutil"
 )
 
 func TestValidateBlock1(t *testing.T) {
