+++ protocol/vm/vmutil/builder.go
