+++ protocol/vm/vmutil/builder_test.go
