+++ protocol/vm/vmutil/script.go
