+++ protocol/vm/vmutil/script_test.go
